Hotfix 1.0.0

See merge request project/rcdc/cbb/rcdc-terminal-module!25

+++ rcdc-terminal-module-impl/src/main/java/com/ruijie/rcos/rcdc/terminal/module/impl/entity/TerminalEntity.java
@@ -57,9 +57,9 @@ public class TerminalEntity {
 
     private String cpuType;
 
-    private Integer memorySize;
+    private Long memorySize;
 
-    private Integer diskSize;
+    private Long diskSize;
 
     private String terminalOsType;
 
@@ -191,19 +191,19 @@ public class TerminalEntity {
         this.cpuType = cpuType;
     }
 
-    public Integer getMemorySize() {
+    public Long getMemorySize() {
         return memorySize;
     }
 
-    public void setMemorySize(Integer memorySize) {
+    public void setMemorySize(Long memorySize) {
         this.memorySize = memorySize;
     }
 
-    public Integer getDiskSize() {
+    public Long getDiskSize() {
         return diskSize;
     }
 
-    public void setDiskSize(Integer diskSize) {
+    public void setDiskSize(Long diskSize) {
         this.diskSize = diskSize;
     }
